Reorg twizzler-device::request.
